NLP Sentiment Analysis Using Bernoulli's Distribution.
Problem Background

Business Understanding
With a growing trend towards digitisation and prevalence of mobile phones and internet access, more consumers have an online presence and their opinions hold a good value for any product-based company, especially so for the B2C businesses. The industries are trying to fine-tune their strategies to suit the consumer needs, as the consumers leave some hints of their choices during their online presence. 

 

In this capstone project, you will be solving a problem in the mobile phone industry of the US, one of the major smartphones markets in the world. In the next video, Tanushree will talk about some of the statistics in the US phone market and how the businesses benefit from analysing the market data.